refactor(layout): move GuardSpec to phase_state to break the guard/routing cycle
`routing/invariants.py` imported `GuardSpec` from `layout/phases/guards.py`,
so guards.py could not import the routing layer at module scope and instead
re-imported it from inside function bodies: 73 of its 106 function-local
`nf_metro` imports were the routing modules, the same handful over and over.

`GuardSpec` now lives in `layout/phase_state.py`, which has no `nf_metro`
imports of its own and already holds the stage vocabulary that
`GuardSpec.first_valid_stage` names. That single edge was the whole cycle:
guards.py now imports `routing`, `routing.common`, `routing.context`,
`routing.core`, `routing.invariants`, `routing.normalize`, `routing.rail` and
`routing.reversal` at module scope, and the 73 duplicated function-local
imports are gone.

The 33 function-local imports that remain are unrelated to this cycle: 9 sit
on genuine cycles (`phases.fan_bundles`, `engine` and `render.svg` all import
guards.py), 7 duplicate an existing module-level import, and 17 reach modules
outside this change's boundary.

Relocation only. No signature, logic or ordering change; all 333 shipped
`examples/**/*.mmd` render byte-identically, with no error on either side.
