+++ src/nf_metro/layout/phases/guards.py
@@ -6,7 +6,6 @@ import math
 import warnings
 from collections import defaultdict
 from collections.abc import Callable, Iterator
-from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, NamedTuple
 
 from nf_metro.errors import NfMetroError
@@ -46,6 +45,7 @@ from nf_metro.layout.geometry import (
     segment_intersects_bbox,
 )
 from nf_metro.layout.pass_metrics import icon_half_height_approx, station_radius_approx
+from nf_metro.layout.phase_state import GuardSpec
 from nf_metro.layout.phases._common import (
     _bbox_cols_overlap,
     _canvas_width,
@@ -99,6 +99,56 @@ from nf_metro.layout.phases.spacing import (
     _placed_name_label_station_ids,
     _residual_label_overlaps,
 )
+from nf_metro.layout.routing import (
+    compute_station_offsets,
+    observe_route_edges,
+    route_edges,
+    route_edges_centred,
+)
+from nf_metro.layout.routing.common import (
+    RoutedPath,
+    _center_inter_row_channel,
+    _inter_row_band_fits,
+    apply_route_offsets,
+    endpoint_port_xs,
+    resolve_section,
+    row_bottom_edge,
+    row_top_edge,
+    section_exists_above_row,
+)
+from nf_metro.layout.routing.context import _build_routing_context
+from nf_metro.layout.routing.core import (
+    _h_segment_crosses_other_section,
+    compute_junction_fan_info,
+    route_edges_for_placement_guards,
+)
+from nf_metro.layout.routing.invariants import (
+    bypass_horizontal_targets,
+    check_convergence_shallow_feeder_concentric,
+    check_exit_inherits_entry_bundle_order,
+    check_fan_merge_no_partition_crossing,
+    check_fanout_tail_join,
+    check_merge_port_approach_side,
+    check_merge_port_outgoing_side_preserved,
+    check_no_distinct_line_fanout_crossing,
+    check_no_dogleg_crosses_exempt_trunk,
+    check_no_split_same_line_fanout_descents,
+    check_partial_branch_offset_gaps,
+    check_perp_entry_boundary_consistent,
+    check_perp_exit_over_leadin_clears_only_spanned_sections,
+    check_right_entry_corridor_descent_no_jog,
+    check_right_entry_drop_in_when_clear,
+    check_stacked_elbow_clearance,
+    check_stacked_split_no_line_recrossing,
+    check_station_bundle_contiguous_at_fan_port,
+    check_tb_exit_corner_preserves_column_order,
+    check_trunk_continuation_drops_straight,
+    classify_merge_port_feeders,
+    distinct_offset_levels,
+)
+from nf_metro.layout.routing.normalize import _suboptimal_trunk_bands
+from nf_metro.layout.routing.rail import _line_rail_y
+from nf_metro.layout.routing.reversal import tb_positive_fan_sections
 from nf_metro.parser.model import (
     LineSpread,
     MetroGraph,
@@ -125,7 +175,6 @@ if TYPE_CHECKING:
         RealisedRouteReservation,
         RouteReservation,
     )
-    from nf_metro.layout.routing.common import RoutedPath
 
     class _HasMessage(Protocol):
         def message(self) -> str: ...
@@ -621,8 +670,6 @@ def _guard_ports_clear_unanchored_box_edges(
     of the settled layout, not of every stage boundary.
     """
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     for pid, port in graph.ports.items():
         st = graph.stations.get(pid)
@@ -1560,8 +1607,6 @@ def _guard_corridor_fed_solo_rides_trunk(
     downstream multi-line section, so they are not required to ride offset 0.
     """
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     for sec_id, pid, line_id in iter_corridor_fed_solo_entries(graph, SAME_Y_TOLERANCE):
         port_off = offsets.get((pid, line_id), 0.0)
@@ -2288,7 +2333,6 @@ def _guard_no_stacked_elbow_graze(
     within ``BUNDLE_TO_BUNDLE_CLEARANCE`` of each other their opposing elbows
     overlap and the lines graze instead of reading as distinct streams.
     """
-    from nf_metro.layout.routing.invariants import check_stacked_elbow_clearance
 
     _raise_on_first_violation(
         graph, phase, check_stacked_elbow_clearance, offsets, routes
@@ -2308,8 +2352,6 @@ def _guard_no_station_overlap(
     once a candidate's left edge passes the current bbox's right edge.
     """
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     boxes: list[tuple[str, tuple[float, float, float, float]]] = []
     for sid in graph.stations:
@@ -2371,15 +2413,10 @@ def _guard_no_line_crosses_non_consumer(
     sharing its trunk-Y row with a busier sibling whose inbound
     bundle traverses the sparse consumer's column.
     """
-    from nf_metro.layout.routing.common import apply_route_offsets
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     if routes is None:
-        from nf_metro.layout.routing import route_edges
-
         try:
             routes = route_edges(graph, station_offsets=offsets)
         except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2444,17 +2481,12 @@ def _guard_no_line_crosses_file_icon(
     one belonging to a line the icon's station also carries, since a
     different edge of that line is still raking the artefact.
     """
-    from nf_metro.layout.routing.common import apply_route_offsets
     from nf_metro.render.svg import _icon_obstacles_by_station
     from nf_metro.themes import THEMES
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     if routes is None:
-        from nf_metro.layout.routing import route_edges
-
         try:
             routes = route_edges(graph, station_offsets=offsets)
         except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2564,15 +2596,10 @@ def iter_opposing_line_overlaps(
     lines sharing a channel are carried on their own offset slots, not on the
     same track.
     """
-    from nf_metro.layout.routing.common import apply_route_offsets
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     if routes is None:
-        from nf_metro.layout.routing import route_edges
-
         try:
             routes = route_edges(graph, station_offsets=offsets)
         except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2669,19 +2696,14 @@ def iter_line_label_strikes(
         place_labels,
         segment_strikes_label,
     )
-    from nf_metro.layout.routing.common import apply_route_offsets
     from nf_metro.render.svg import _compute_icon_obstacles
     from nf_metro.themes import THEMES
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     with _restoring_layout_geometry(graph):
         if routes is None:
-            from nf_metro.layout.routing import route_edges_centred
-
             try:
                 routes = route_edges_centred(graph, station_offsets=offsets)
             except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2834,14 +2856,10 @@ def _guard_no_wrapped_label_trunk_strike(
     from nf_metro.themes import THEMES
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     with _restoring_layout_geometry(graph):
         if routes is None:
-            from nf_metro.layout.routing import route_edges_centred
-
             try:
                 routes = route_edges_centred(graph, station_offsets=offsets)
             except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2879,7 +2897,6 @@ def _guard_off_track_output_clears_non_producer(
     closes the gap by checking the output route against same-section trunk
     markers regardless of line membership, exempting only the producer.
     """
-    from nf_metro.layout.routing.common import apply_route_offsets
 
     producer_of = {
         off_id: anchor_id
@@ -2890,12 +2907,8 @@ def _guard_off_track_output_clears_non_producer(
         return
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     if routes is None:
-        from nf_metro.layout.routing import route_edges
-
         try:
             routes = route_edges(graph, station_offsets=offsets)
         except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -2947,8 +2960,6 @@ def _guard_row_trunk_cy_consistent(
     line sets (e.g. parallel sub-rows on a row-spanner) don't trigger.
     """
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     rows: dict[int, list[Section]] = {}
@@ -3080,8 +3091,6 @@ def _guard_inter_section_routes_in_row_band(
     (bypass clearance + bundle nest + diagonal corner approach).
     """
     if offsets is None or routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         if routes is None:
@@ -3147,10 +3156,6 @@ def _guard_topmost_row_top_entry_hugs_section(
     route's-width above the section edge. A deeper climb drives the line
     through the title text.
     """
-    from nf_metro.layout.routing.common import (
-        row_top_edge,
-        section_exists_above_row,
-    )
 
     routes = _ensure_routes(graph, routes)
 
@@ -3233,7 +3238,6 @@ def _ensure_routes(
     """
     if routes is not None:
         return routes
-    from nf_metro.layout.routing import compute_station_offsets, route_edges
 
     return route_edges(graph, station_offsets=compute_station_offsets(graph))
 
@@ -3275,7 +3279,6 @@ def _inter_section_backtrack_legs(
     column legitimately wrap and are skipped, as are TB folds and
     same-column routes.
     """
-    from nf_metro.layout.routing.common import resolve_section
 
     for rp in routes:
         if not rp.is_inter_section:
@@ -3324,7 +3327,6 @@ def _guard_inter_section_route_no_backtrack(
     skipped, as are ``normalize_exempt`` wrap legs, TB folds, and same-column
     routes.
     """
-    from nf_metro.layout.routing.common import resolve_section
 
     routes = _ensure_routes(graph, routes)
 
@@ -3429,8 +3431,6 @@ def _guard_fan_bundles_coincide_or_separate(
     as a smeared partial overlap rather than one bundle or two separated
     bundles.
     """
-    from nf_metro.layout.routing import compute_station_offsets, route_edges
-    from nf_metro.layout.routing.core import compute_junction_fan_info
 
     if offsets is None:
         offsets = compute_station_offsets(graph)
@@ -3562,7 +3562,6 @@ def _guard_port_legs_meet_rail_pill_on_own_rail(
     """
     if not graph.has_rail_sections:
         return
-    from nf_metro.layout.routing.rail import _line_rail_y
 
     routes = _ensure_routes(graph, routes)
 
@@ -3975,14 +3974,6 @@ def _guard_no_artefactual_counter_flow(
     guard fires only when the with-flow gap channel above the target row was
     genuinely free for the run's X-span yet went unused.
     """
-    from nf_metro.layout.routing.common import (
-        _center_inter_row_channel,
-        _inter_row_band_fits,
-        resolve_section,
-        row_bottom_edge,
-        row_top_edge,
-    )
-    from nf_metro.layout.routing.core import _h_segment_crosses_other_section
 
     routes = _ensure_routes(graph, routes)
 
@@ -4076,8 +4067,6 @@ def _guard_serpentine_no_backtrack(
     """
     routes = _ensure_routes(graph, routes)
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     for sid, against, limit, section in iter_serpentine_backtracks(
@@ -4106,7 +4095,6 @@ def _guard_inter_row_run_clearance(
     (``_wrap_bundle_row_minimums``) reserves the space; this guard fails
     loudly if a layout change ever lets the run creep back against the box.
     """
-    from nf_metro.layout.routing.common import resolve_section
 
     routes = _ensure_routes(graph, routes)
 
@@ -4165,8 +4153,6 @@ def _guard_trunk_bands_crossing_optimal(
     leaves a band in such an avoidable-crossing order.
     """
     from nf_metro.layout.constants import CURVE_RADIUS, DIAGONAL_RUN
-    from nf_metro.layout.routing.context import _build_routing_context
-    from nf_metro.layout.routing.normalize import _suboptimal_trunk_bands
 
     routes = _ensure_routes(graph, routes)
     ctx = _build_routing_context(graph, DIAGONAL_RUN, CURVE_RADIUS, offsets)
@@ -4197,7 +4183,6 @@ def _guard_inter_section_descent_edge_clearance(
     in :func:`_route_l_shape` pushes such channels outward; this guard
     fails loudly if a future change lets one creep back against an edge.
     """
-    from nf_metro.layout.routing.common import endpoint_port_xs
 
     routes = _ensure_routes(graph, routes)
 
@@ -4253,9 +4238,6 @@ def _guard_tb_exit_corner_column_order(
     the column swaps two lines' X and renders a crossing through the feeder
     station marker.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_tb_exit_corner_preserves_column_order,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_tb_exit_corner_preserves_column_order, offsets, routes
@@ -4276,9 +4258,6 @@ def _guard_no_split_same_line_fanout_descents(
     ride one fused trunk rather than open at distinct Xs, which would peel the
     farther-reaching branch onto the inside of the nearer one and cross it.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_no_split_same_line_fanout_descents,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_no_split_same_line_fanout_descents, offsets, routes
@@ -4299,9 +4278,6 @@ def _guard_no_distinct_line_fanout_crossing(
     descend as one unit and split only where each line turns into its target,
     never crossing a mate's run on the way down.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_no_distinct_line_fanout_crossing,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_no_distinct_line_fanout_crossing, offsets, routes
@@ -4323,10 +4299,6 @@ def _guard_fan_merge_no_partition_crossing(
     horizontal split fed by a stacked half-turn, a line pair may not cross and
     recross within the consumer section.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_fan_merge_no_partition_crossing,
-        check_stacked_split_no_line_recrossing,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_fan_merge_no_partition_crossing, offsets, routes
@@ -4350,9 +4322,6 @@ def _guard_trunk_continuation_drops_straight(
     peeling past a sibling, or a collinear feeder into a terminal merge -- it
     must run straight rather than jog by one step off the trunk.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_trunk_continuation_drops_straight,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_trunk_continuation_drops_straight, offsets, routes
@@ -4373,9 +4342,6 @@ def _guard_no_dogleg_crosses_exempt_trunk(
     on the side that keeps the two parallel, never the side whose riser pierces
     the exempt run and crosses it twice.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_no_dogleg_crosses_exempt_trunk,
-    )
 
     _raise_on_first_violation(
         graph, phase, check_no_dogleg_crosses_exempt_trunk, offsets, routes
@@ -4394,12 +4360,8 @@ def _raise_on_first_violation(
 ) -> None:
     """Run a route *check* and raise ``PhaseInvariantError`` on its first hit."""
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
     if routes is None:
-        from nf_metro.layout.routing import route_edges
-
         try:
             routes = route_edges(graph, station_offsets=offsets)
         except Exception:  # noqa: BLE001 - routing failure surfaces elsewhere
@@ -4429,11 +4391,8 @@ def _guard_merge_port_approach_side(
     :func:`nf_metro.layout.routing.invariants.check_merge_port_approach_side`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import check_merge_port_approach_side
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     violations = check_merge_port_approach_side(graph, offsets)
@@ -4458,13 +4417,8 @@ def _guard_station_bundle_contiguous_at_fan_port(
     :func:`nf_metro.layout.routing.invariants.check_station_bundle_contiguous_at_fan_port`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_station_bundle_contiguous_at_fan_port,
-    )
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     violations = check_station_bundle_contiguous_at_fan_port(graph, offsets)
@@ -4490,13 +4444,8 @@ def _guard_convergence_shallow_feeder_concentric(
     :func:`nf_metro.layout.routing.invariants.check_convergence_shallow_feeder_concentric`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_convergence_shallow_feeder_concentric,
-    )
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     messages = check_convergence_shallow_feeder_concentric(graph, offsets)
@@ -4520,13 +4469,8 @@ def _guard_merge_port_outgoing_side_preserved(
     :func:`nf_metro.layout.routing.invariants.check_merge_port_outgoing_side_preserved`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_merge_port_outgoing_side_preserved,
-    )
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     violations = check_merge_port_outgoing_side_preserved(graph, offsets)
@@ -4551,13 +4495,8 @@ def _guard_exit_inherits_entry_bundle_order(
     :func:`nf_metro.layout.routing.invariants.check_exit_inherits_entry_bundle_order`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_exit_inherits_entry_bundle_order,
-    )
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     violations = check_exit_inherits_entry_bundle_order(graph, offsets)
@@ -4581,15 +4520,8 @@ def _guard_bypass_port_no_slot_gaps(
     inflates ``max_horiz`` and pushes perpendicular feeders into outer slots,
     leaving empty slots between the horizontal band and the feeders.
     """
-    from nf_metro.layout.routing.invariants import (
-        bypass_horizontal_targets,
-        classify_merge_port_feeders,
-        distinct_offset_levels,
-    )
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     for port_id in graph.ports:
@@ -4630,11 +4562,8 @@ def _guard_partial_branch_offset_gaps(
     :func:`nf_metro.layout.routing.invariants.check_partial_branch_offset_gaps`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import check_partial_branch_offset_gaps
 
     if offsets is None:
-        from nf_metro.layout.routing import compute_station_offsets
-
         offsets = compute_station_offsets(graph)
 
     violations = check_partial_branch_offset_gaps(graph, offsets)
@@ -4660,11 +4589,8 @@ def _guard_fanout_tail_join(
     See :func:`nf_metro.layout.routing.invariants.check_fanout_tail_join`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import check_fanout_tail_join
 
     if routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         try:
@@ -4695,13 +4621,8 @@ def _guard_perp_entry_boundary_consistent(
     :func:`nf_metro.layout.routing.invariants.check_perp_entry_boundary_consistent`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_perp_entry_boundary_consistent,
-    )
 
     if routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         try:
@@ -4732,13 +4653,8 @@ def _guard_perp_exit_over_leadin_no_overdip(
     :func:`nf_metro.layout.routing.invariants.check_perp_exit_over_leadin_clears_only_spanned_sections`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_perp_exit_over_leadin_clears_only_spanned_sections,
-    )
 
     if routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         try:
@@ -4768,13 +4684,8 @@ def _guard_right_entry_drop_in_when_clear(
     :func:`nf_metro.layout.routing.invariants.check_right_entry_drop_in_when_clear`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_right_entry_drop_in_when_clear,
-    )
 
     if routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         try:
@@ -4804,13 +4715,8 @@ def _guard_right_entry_corridor_descent_no_jog(
     :func:`nf_metro.layout.routing.invariants.check_right_entry_corridor_descent_no_jog`
     for the semantic definition.
     """
-    from nf_metro.layout.routing.invariants import (
-        check_right_entry_corridor_descent_no_jog,
-    )
 
     if routes is None:
-        from nf_metro.layout.routing import compute_station_offsets, route_edges
-
         if offsets is None:
             offsets = compute_station_offsets(graph)
         try:
@@ -4940,7 +4846,6 @@ def _guard_fanout_junction_resolves_upstream(graph: MetroGraph, phase: str) -> N
     section would leave routing without a grid column/row for it and silently
     misplace the fanned bundle.
     """
-    from nf_metro.layout.routing.common import resolve_section
 
     topology = build_route_topology_query(graph)
     junction_ids = (
@@ -5347,7 +5252,6 @@ def _guard_planned_fan_frame_realised(
         vertical_fan_label_lane_pitch,
     )
     from nf_metro.layout.route_plan import FanAppearancePolicy, fan_lane_seat_keys
-    from nf_metro.layout.routing.reversal import tb_positive_fan_sections
 
     contract_fault = _fan_plan_contract_fault(graph)
     if contract_fault is not None:
@@ -5569,45 +5473,6 @@ def _guard_canvas_margin_settled(
     if graph.strict and not graph.permissive:
         raise LayoutInvariantError(msg)
     warnings.warn(msg, category=PermissiveGuardWarning, stacklevel=2)
-
-
-@dataclass(frozen=True)
-class GuardSpec:
-    """One ``validate=True`` guard, with the dispatch + classification data
-    that used to be scattered across hand-written call sites and the
-    ``_BISECTION_FIRST_VALID`` table.
-
-    ``fn`` is the guard function; every guard takes ``(graph, phase)`` and the
-    optional keyword inputs named in ``needs`` (a subset of ``offsets``,
-    ``routes``, ``section_y_gap``, ``section_y_padding``).  The dispatcher
-    passes exactly those keywords, so heterogeneous signatures need no
-    wrapping.
-
-    ``bisection_safe`` guards run at every Pass C checkpoint (gated by
-    ``first_valid_stage``, the earliest checkpoint at which their invariant
-    holds) as well as at the closing ``after final`` boundary; the rest run
-    only at ``after final``.  ``tier`` is the cost-tier classification
-    (``docs/dev/guard_tiers.md``).
-
-    ``issue_pin`` is the tuple of ``#NNN`` issues a guard was born from; it
-    keeps the regression trail as data so consolidating or renaming a guard
-    cannot let the original bug be silently re-filed.  ``narrow_reason`` states
-    why a guard pinned to an issue stays scoped to its case rather than being
-    folded into a broader geometric property -- a required field for any
-    issue-pinned guard (``test_issue_pinned_guards_document_why_they_are_narrow``).
-    """
-
-    fn: Callable[..., Any]
-    tier: str
-    needs: frozenset[str] = field(default_factory=frozenset)
-    bisection_safe: bool = False
-    first_valid_stage: str | None = None
-    issue_pin: tuple[str, ...] = ()
-    narrow_reason: str | None = None
-
-    @property
-    def name(self) -> str:
-        return self.fn.__name__
 
 
 # The single ordered source of truth for the ``validate=True`` guard
@@ -6282,8 +6147,6 @@ def _ensure_pass_c_inputs(
     render chokepoint, which receives the settled routes. Guards without a
     route dependency continue to inspect the final layout state.
     """
-    from nf_metro.layout.routing import compute_station_offsets, observe_route_edges
-    from nf_metro.layout.routing.core import route_edges_for_placement_guards
 
     if offsets is None:
         offsets = compute_station_offsets(graph)
